refactor: 액션바 컴포넌트 분리
