+++ src/app/workspace/[id]/page.tsx
@@ -1,11 +1,10 @@
 // 작업공간
-import ActionBar from '@components/action-bar/ActionBar'
+import WorkspaceActionBar from './components/workspace-action-bar/WorkspaceActionBar'
 
 export default function Page() {
   return (
     <div style={{ height: '100vh' }}>
-      <ActionBar usage="workspace" />
-      <ActionBar usage="planner" />
+      <WorkspaceActionBar />
     </div>
   )
 }
